requirements.txt
arxiv
pyyaml
feedparser
requests
